% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bailey_Clutter_B0.R
\name{Bailey_Clutter.B0}
\alias{Bailey_Clutter.B0}
\title{Modelo anamórfico \eqn{\beta_0} de Bailey y Clutter (1974)}
\usage{
Bailey_Clutter.B0(B1, B2, Y0, E0, E1)
}
\arguments{
\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica (ADA, por sus siglas en inglés). El parámetro de la asíntota (\eqn{\beta_0})
se relacionó con el sitio.
}
\details{
El modelo ADA con valor específico del sitio \eqn{\beta_0} es:
    \deqn{y_1 = exp \left( (log(Y_0)-\beta_1\times E_0^{\beta_2})+\beta_1 \times E_1^{\beta_2} \right)}
}
\references{
Bailey & Clutter (1974). Base-age invariant polymorphic site curves.
    Forest Science. 20(2):155-159.
}
\seealso{
\code{\link{Bailey_Clutter}}
}
\author{
Abel Joseph Hernández-Martínez
}
