% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bailey_Clutter_G1.R
\name{Bailey_Clutter_G1}
\alias{Bailey_Clutter_G1}
\title{Modelo polimórfico complejo (GADA 1) de Bailey y Clutter (1974)}
\usage{
Bailey_Clutter_G1(B0, B1, B2, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
    Algebraica Generalizada (GADA, por sus siglas en inglés). La ecuación original se reparametrizó
    \eqn{y =  exp \left( \alpha_0 + \alpha_1 E^{\beta_2}  \right) }, donde:
    \deqn{\alpha_0 = \beta_0 + \chi}
    \deqn{\alpha_1 = \beta_1  \times \chi}
    \deqn{\alpha_2 = \beta_2}
}
\details{
El modelo GADA:
    \deqn{Y_1 = exp \left( \beta_0 + \chi_0 +\beta_1 \chi_0 E_0^{\beta_2} \right)}
    donde:
    \deqn{\chi_0 = \frac{log(Y_0)-\beta_0}{1-\beta_1 \times E_0^{\beta_2}}}
}
\references{
Özçelik et al  (2019). Modeling dominant height growth of cedar
    (\emph{Cedrus libani} A. Rich) stands in Turkey. Forest Science, 65(6),
    725-733.
}
\seealso{
\code{\link{Bailey_Clutter}}
}
\author{
Abel Joseph Hernández-Martínez
}
