% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hossfeld_IV_exp_B2.R
\name{Hossfeld_IV_exp_B2}
\alias{Hossfeld_IV_exp_B2}
\title{Modelo polimórfico \eqn{\beta_2} de Hossfeld IV exponencial (1882)}
\usage{
Hossfeld_IV_exp_B2(B0, B1, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica (ADA, por sus siglas en inglés). El parámetro de la
tasa de cambio (\eqn{\beta_2}) se relacionó con el sitio.
}
\details{
El modelo ADA con valor específico del sitio \eqn{\beta_2}:
    \deqn{Y_1 = \frac{\beta_0}{1+exp(\beta_1) \times exp \left( \left( \frac{log \left( \frac{\beta_0 - Y_0}{Y_0 \times exp(\beta_1)} \right)}{log(E_0)} \right) \times log(E_1) \right)}}
}
\references{
Hernández-Cuevas et al. (2018). Modelos de crecimiento en altura
dominante e índices de sitio para \emph{Pinus ayacahuite} Ehren.
Agrociencia 52:437-453.
}
\seealso{
\code{\link{Hossfeld_IV_exp}}
}
\author{
Abel Joseph Hernández-Martínez
}
