% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hossfeld_IV_exp_G1.R
\name{Hossfeld_IV_exp_G1}
\alias{Hossfeld_IV_exp_G1}
\title{Modelo polimórfico complejo (GADA 1) de Hossfeld IV exponencial (1822)}
\usage{
Hossfeld_IV_exp_G1(B0, B1, B2, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
    Algebraica Generalizada (GADA, por sus siglas en inglés). La ecuación original se reparametrizó:
    \deqn{y = \frac{\alpha_0}{a+exp(\alpha_1) \times exp(-\alpha_2 \times log(E))} } donde:
    \deqn{\alpha_0 = \beta_0 + \chi}
    \deqn{\alpha_1 = \frac{\beta_1}{\chi}}
    \deqn{\alpha_2 = \beta_2}
}
\details{
El modelo GADA:
    \deqn{Y_1 = \frac{\beta_1 + \chi_0}{1 + \frac{\beta_1}{\chi_0} \times E_1^{-\beta_2} }}
    donde:
    \deqn{\chi_0 = \frac{Y_0 - \beta_0 + \sqrt{(Y_0-\beta_0)^2 - 4 \times \beta_1 \times Y_0 \times E_0^{-\beta_2}}}{2}}
}
\references{
Hernández-Cuevas et al. (2018). Modelos de crecimiento en altura
dominante e índices de sitio para \emph{Pinus ayacahuite} Ehren.
Agrociencia 52:437-453.
}
\seealso{
\code{\link{Hossfeld_IV_exp}}
}
\author{
Abel Joseph Hernández-Martínez
}
