% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Korf_B0.R
\name{Korf_B0}
\alias{Korf_B0}
\title{Modelo anamórfico \eqn{\beta_0} de Korf (1939)}
\usage{
Korf_B0(B1, B2, Y0, E0, E1)
}
\arguments{
\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica (ADA, por sus siglas en inglés). El parámetro de la asíntota (\eqn{\beta_0})
se relacionó con el sitio.
}
\details{
El modelo ADA con valor específico del sitio \eqn{\beta_0}:
    \deqn{Y_1 = Y_0 \times \left( \frac{exp \left(-\beta_1 \times E_1^{-\beta_2} \right)}{exp \left(-\beta_1 \times E_0^{-\beta_2} \right)} \right)}
}
\seealso{
\code{\link{Korf}}
}
\author{
Abel Joseph Hernández-Martínez
}
