% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Korf_B2_2.R
\name{Korf_B2_2}
\alias{Korf_B2_2}
\title{Modelo polimórfico \eqn{\beta_2} segunda opción de Korf (1939)}
\usage{
Korf_B2_2(B0, B1, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica (ADA, por sus siglas en inglés). El parámetro \eqn{\beta_2}
se relacionó con el sitio.
}
\details{
El modelo ADA con valor específico del sitio \eqn{\beta_2}:
    \deqn{Y_1 = \beta_0 \times exp \left( \frac{-\beta_1} {E_1^{\left( \frac{log \left( \frac{-\beta_1}{log \left( \frac{Y_0}{\beta_0} \right)} \right)}{log(E_0)} \right)} }\right)}
}
\seealso{
\code{\link{Korf}}
}
\author{
Abel Joseph Hernández-Martínez
}
