% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Korf_G1.R
\name{Korf_G1}
\alias{Korf_G1}
\title{Modelo polimórfico complejo (GADA 1) de Korf (1939)}
\usage{
Korf_G1(B1, B2, Y0, E0, E1)
}
\arguments{
\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica Generalizada (GADA, por sus siglas en inglés). La ecuación original se reparametrizó:
\deqn{y = \alpha_0 \times exp \left( -\alpha_1 \times E^{-\alpha_2} \right) } donde:
\deqn{\alpha_0 = exp(\chi)}
\deqn{\alpha_1 = \beta_1 + \frac{1}{\chi}}
\deqn{\alpha_2 = \beta_2}
}
\details{
El modelo GADA:
    \deqn{Y_1 = exp(\chi_0) \times exp \left( \frac{-1}{E_1^{\beta_2}} \right)^{\left( \beta_1 + \frac{1}{\chi_0} \right)}}
    donde:
    \deqn{\chi_0 = 0.5 \times (-L_0 + R_0)}
    \deqn{R_0 = \sqrt{L_0^2 - 4 \times log \left( exp \left( \frac{-1}{E_0^{\beta_2}} \right) \right) }}
    \deqn{L_0 =  \beta_1 \times \left( \frac{-1}{E_0^{\beta_2}} \right) - log(Y_0)}
}
\references{
Tamarit-Urias et al. (2014). Ecuaciones dinámicas de índice de
   sitio para \emph{Tectona grandis} en Campeche, México. Agrociencia 48: 225-238.
}
\seealso{
\code{\link{Korf}}
}
\author{
Abel Joseph Hernández-Martínez
}
