% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Korf_G2.R
\name{Korf_G2}
\alias{Korf_G2}
\title{Modelo polimórfico complejo (GADA 2) de Korf (1939)}
\usage{
Korf_G2(B0, B1, B2, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica Generalizada (GADA, por sus siglas en inglés). La ecuación original se reparametrizó:
\deqn{y = \alpha_0 \times exp \left( -\alpha_1 \times E^{-\alpha_2} \right) } donde:
\deqn{\alpha_0 = exp(\chi)}
\deqn{\alpha_1 = \beta_0 + \frac{\beta_1}{\chi}}
\deqn{\alpha_2 = \beta_2}
}
\details{
El modelo GADA:
    \deqn{Y_1 = exp(\chi_0) \times exp \left( \left(-\beta_0 + \frac{\beta_1}{\chi_0} \right) \times E_1^{-\beta_2} \right)}
    donde:
    \deqn{\chi_0 = 0.5 \times \beta_0 \times E_0^{-\beta_2} + log(Y_0) + R_0}
    \deqn{R_0 = \sqrt{\left(\beta_0 \times E_0^{-\beta_2} + log(Y_0) \right)^{2} + 4 \times \beta_1 \times E_0^{-\beta_2}}}
}
\references{
Tamarit-Urias et al. (2014). Ecuaciones dinámicas de índice de
   sitio para \emph{Tectona grandis} en Campeche, México. Agrociencia 48: 225-238.
}
\seealso{
\code{\link{Korf}}
}
\author{
Abel Joseph Hernández-Martínez
}
