% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levakovic_II_B1.R
\name{Levakovic_II_B1}
\alias{Levakovic_II_B1}
\title{Modelo polimórfico \eqn{\beta_1} de Levakovic II (1935)}
\usage{
Levakovic_II_B1(B0, B2, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica (ADA, por sus siglas en inglés) con el parámetro de la
tasa de crecimiento.
}
\details{
El modelo ADA con valor específico del sitio \eqn{\beta_1} es:
    \deqn{Y_1 = \beta_0 \times \left( E_1 \times \left( \left( \frac{E_0} {exp \left( \frac{log(Y_0/\beta_0)}{\beta_2} \right)}-E_0 \right)+E_1 \right)^{-1} \right)^{\beta_2}}
}
\references{
Hernández-Cuevas et al. (2018). Modelos de crecimiento en altura
dominante e índices de sitio para \emph{Pinus ayacahuite} Ehren.
Agrociencia 52:437-453.
}
\seealso{
\code{\link{Levakovic_II}}
}
\author{
Abel Joseph Hernández-Martínez
}
