% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levakovic_II_G1.R
\name{Levakovic_II_G1}
\alias{Levakovic_II_G1}
\title{Modelo polimórfico complejo (GADA 1) de Levakovic II (1935)}
\usage{
Levakovic_II_G1(B1, B2, Y0, E0, E1)
}
\arguments{
\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica Generalizada (GADA, por sus siglas en inglés). La ecuación original se reparametrizó
\eqn{y = \alpha_0 \times \left( \frac{E}{\alpha_1+E} \right)^{\alpha_2} }, donde:
\deqn{\alpha_0 = exp(\chi)}
\deqn{\alpha_1 = \beta_1}
\deqn{\alpha_2 = \frac{\beta_2}{\chi}}
}
\details{
El modelo GADA:
    \deqn{Y_1 = exp(\chi_0) \times \left( \frac{E_1}{\beta_1+E_1} \right)^{\frac{\beta_2}{\chi_0} } }
    donde:
    \deqn{\chi_0 = \frac{log(Y_0) + \sqrt{log(Y_0)^2 - 4 \times \beta_1 \times \left( \frac{E_1}{\beta_1+E_1} \right)} }{2}}
}
\references{
Vera-Vélez et al. (2013). Efecto de variables climáticas y litológicas
    sobre el crecimiento en altura dominante en masas de \emph{Pinus nigra} Arn.
}
\seealso{
\code{\link{Levakovic_II}}
}
\author{
Abel Joseph Hernández-Martínez
}
