% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Richards_B1.R
\name{Richards_B1}
\alias{Richards_B1}
\title{Modelo polimórfico \eqn{\beta_1} de Chapman-Richards (1959)}
\usage{
Richards_B1(B0, B2, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica (ADA, por sus siglas en inglés) con el parámetro de la
tasa de crecimiento.
}
\details{
El modelo ADA con valor específico del sitio \eqn{\beta_1} es:
    \deqn{y_1 = \beta_0 \times \left(1-\left(1-\left(\frac{Y_0}{\beta_0}\right)^{(\frac{1}{\beta_2})}\right)^{(\frac{E_1}{E_0})}\right)^{B2}}
}
\references{
Hernández-Cuevas et al. (2018). Modelos de crecimiento en altura
dominante e índices de sitio para \emph{Pinus ayacahuite} Ehren.
Agrociencia 52:437-453.
}
\seealso{
\code{\link{Richards}}
}
\author{
Abel Joseph Hernández-Martínez
}
