% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Richards_G2.R
\name{Richards_G2}
\alias{Richards_G2}
\title{Modelo polimórfico complejo (GADA 2) de Chapman-Richards (1935)}
\usage{
Richards_G2(B1, B2, Y0, E0, E1)
}
\arguments{
\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
    Algebraica Generalizada (GADA, por sus siglas en inglés). La ecuación original se reparametrizó:
    \deqn{y = \alpha_0 \times \left( 1- exp(-\alpha_1 \times E) \right)^{\alpha_2}} donde:
    \deqn{\alpha_0 = \exp(\chi)}
    \deqn{\alpha_1 = \beta_1}
    \deqn{\alpha_2 = \frac{\beta_2}{\chi}}
}
\details{
El modelo GADA:
    \deqn{Y_1 = exp(\chi_0) \times \left( 1 - exp(-\beta_1 \times E_1)  \right)^{\left( \frac{\beta_2}{\chi_0} \right)}}
    donde:
    \deqn{\chi_0 = 0.5 \times (log(Y_0) + R_0)}
    \deqn{R_0 = \sqrt{log(Y_0)^{2} - 4 \times \beta_2 \times L_0 }}
    \deqn{L_0 = log(1 - exp(-\beta_1 \times E_0))}
}
\references{
Tamarit-Urias et al. (2014). Ecuaciones dinámicas de índice de
   sitio para \emph{Tectona grandis} en Campeche, México. Agrociencia 48: 225-238.
}
\seealso{
\code{\link{Richards}}
}
\author{
Abel Joseph Hernández-Martínez
}
