% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats_fit.R
\name{Stats_fit}
\alias{Stats_fit}
\title{Estadísticos de bondad ajuste para modelos no lineales}
\usage{
Stats_fit(fit, data)
}
\arguments{
\item{fit}{Un objeto resultado del ajuste de un modelo con \code{\link{nls}}.}

\item{data}{Valores de la variable dependiente a evaluar.}
}
\description{
La función utiliza objetos de ajuste de modelos no lineales
    con la función \code{\link{nls}}.
}
\seealso{
\code{\link{nls}}
}
