% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strand_B0.R
\name{Strand_B0}
\alias{Strand_B0}
\title{Modelo anamórfico \eqn{\beta_0} de Strand generalizado (1964)}
\usage{
Strand_B0(B1, B2, Y0, E0, E1)
}
\arguments{
\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica (ADA, por sus siglas en inglés). El parámetro (\eqn{\beta_0})
se relacionó con el sitio.
}
\details{
El modelo ADA con valor específico del sitio \eqn{\beta_0}:
    \deqn{Y_1 = \left( \frac{E_1}{E_0 \times \left(Y_0^ {\left( \frac{-1}{\beta_2} \right)}- \beta_1 \right)+ \beta_1 \times E_1 }\right)^{\beta_2}}
}
\seealso{
\code{\link{Strand}}
}
\author{
Abel Joseph Hernández-Martínez
}
