% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strand_G1.R
\name{Strand_G1}
\alias{Strand_G1}
\title{Modelo polimórfico complejo (GADA 1) de Strand generalizado (1964)}
\usage{
Strand_G1(B0, B1, B2, Y0, E0, E1)
}
\arguments{
\item{B0}{Valor numérico o parámetro \eqn{\beta_0}.}

\item{B1}{Valor numérico o parámetro \eqn{\beta_1}.}

\item{B2}{Valor numérico o parámetro \eqn{\beta_2}.}

\item{Y0}{Valor del crecimiento inicial.}

\item{E0}{Valor numérico o vector de observaciones de la edad inicial
de los árboles.}

\item{E1}{Valor numérico o vector de observaciones de la edad futura
de los árboles.}
}
\value{
Devuelve el valor del crecimiento estimado del árbol.
}
\description{
La función es el resultado de la aplicación del enfoque de Diferencia
Algebraica Generalizada (GADA, por sus siglas en inglés). La ecuación original se reparametrizó:
\deqn{y = \left( \frac{E}{\alpha_0 + \alpha_1 \times E} \right)^{\alpha_2}} donde:
\deqn{\alpha_0 = exp(\chi)}
\deqn{\alpha_1 = \beta_0 + \beta_1 \times \chi}
\deqn{\alpha_2 = \beta_2}
}
\details{
El modelo GADA:
    \deqn{Y_1 = \left( \frac{E_1}{\chi_0 + (\beta_0 + \beta_1 \times \chi_0) \times E_1 } \right)^{\beta_2}}
    donde:
    \deqn{\chi_0 = \frac{E_0 \times \left( Y_0^{\left( \frac{-1}{\beta_2} \right)} -\beta_0 \right)}{1+ \beta_1 \times E_0}}
}
\references{
Sharma et al. (2011). Modelling dominant height growth from
    national forest inventory individual tree data with short time series and
    large age errors. Forest Ecology and Management 262:2162-2175.
}
\seealso{
\code{\link{Strand}}
}
\author{
Abel Joseph Hernández-Martínez
}
